fix: autoload when register offline table
